refactor(benches): replace RL examples with cartpole_rl and pendulum_rl benches
Convert all nine examples/rl/ training scripts to two multi-algorithm benches
following ADR-0012. cartpole_rl compares DQN, C51, QR-DQN, and PPG against a
random baseline on CartPole; pendulum_rl does the same for PPO, DDPG, TD3, and
SAC on Pendulum. Deferred box2d, locomotion, and bandit random examples also
removed from Cargo.toml; all moved to .scratchpad/random-deferred-examples/.
Adds benches/support/pendulum.rs with shared actor/critic scaffolding.
