--- crates/rlevo/examples/envs/box2d/lunar_lander_discrete_random.rs
@@ -1,40 +1,0 @@
-//! Random-agent smoke test for `LunarLander` (discrete variant).
-
-use rand::RngExt;
-use rand::SeedableRng;
-use rand::rngs::StdRng;
-use rlevo_core::action::DiscreteAction;
-use rlevo_core::environment::{Environment, Snapshot};
-use rlevo_environments::box2d::lunar_lander::{
-    LunarLanderConfig, LunarLanderDiscrete, LunarLanderDiscreteAction,
-};
-
-fn main() {
-    let config = LunarLanderConfig::builder().seed(42).build();
-    let mut env = LunarLanderDiscrete::with_config(config);
-    let mut rng = StdRng::seed_from_u64(0);
-
-    let snap = env.reset().expect("reset failed");
-    println!("reset done, running={}", !snap.is_done());
-
-    let mut total_reward = 0.0f32;
-    for step in 0..500 {
-        let idx = rng.random_range(0..LunarLanderDiscreteAction::ACTION_COUNT);
-        let action = LunarLanderDiscreteAction::from_index(idx);
-        match env.step(action) {
-            Ok(snap) => {
-                let r: f32 = (*snap.reward()).into();
-                total_reward += r;
-                if snap.is_done() {
-                    println!("episode ended at step {step}, total_reward={total_reward:.2}");
-                    return;
-                }
-            }
-            Err(e) => {
-                eprintln!("step error at {step}: {e}");
-                return;
-            }
-        }
-    }
-    println!("truncated after 500 steps, total_reward={total_reward:.2}");
-}
